+++ src/main/java/innohack/gem/entity/feature/CsvFeature.java
@@ -1,4 +1,4 @@
-package innohack.gem.entity.gem.data;
+package innohack.gem.entity.feature;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
@@ -8,7 +8,6 @@ import com.opencsv.CSVReader;
 import com.opencsv.CSVReaderBuilder;
 import com.opencsv.enums.CSVReaderNullFieldIndicator;
 import com.opencsv.exceptions.CsvException;
-
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
