+++ src/main/java/innohack/gem/entity/feature/TikaFeature.java
@@ -1,6 +1,6 @@
-package innohack.gem.entity.gem.data;
+package innohack.gem.entity.feature;
 
-import innohack.gem.entity.gem.util.FeatureExtractorUtil;
+import innohack.gem.entity.feature.common.FeatureExtractorUtil;
 import java.io.File;
 import org.apache.tika.config.TikaConfig;
