+++ src/main/java/innohack/gem/example/TikaExtractor.java
@@ -1,8 +1,8 @@
 package innohack.gem.example;
 
 import innohack.gem.entity.GEMFile;
-import innohack.gem.example.util.FileUtil;
 import innohack.gem.example.tika.TikaUtil;
+import innohack.gem.example.util.FileUtil;
 import java.nio.file.Path;
 import java.util.List;
